Split repositories and place sqlite implementations
