+++ main.py
@@ -7,7 +7,8 @@ from aiogram.enums import ParseMode
 from aiogram.fsm.storage.memory import MemoryStorage
 
 from telegram.controllers.dish_controller import TelegramDishController
-from reposotory.sqllite_dish_repository import SqlliteDishRepository
+from reposotory.sqlite.sqllite_dish_repository import SqlliteDishRepository
+from reposotory.sqlite.sqllite_user_repository import SqlliteUserRepository
 from services.dish_service import DishService
 from services.user_service import UserService
 
@@ -19,11 +20,13 @@ async def main() -> None:
     bot = Bot(token=API_TOKEN, default=DefaultBotProperties(parse_mode=ParseMode.HTML))
     dp = Dispatcher(storage=MemoryStorage())
 
-    repository = SqlliteDishRepository(DB_FILE)
-    await repository.init_db()
+    dish_repo = SqlliteDishRepository(DB_FILE)
+    user_repo = SqlliteUserRepository(DB_FILE)
+    await dish_repo.init_db()
+    await user_repo.init_db()
     admin_logins = json.loads(os.getenv("ADMIN_LOGINS", "[]"))
-    user_service = UserService(repository, bot, admin_logins)
-    dish_service = DishService(repository)
+    user_service = UserService(user_repo, bot, admin_logins)
+    dish_service = DishService(dish_repo)
 
     controller = TelegramDishController(dp, dish_service, user_service)
     controller.register_handlers()
