+++ reposotory/interfaces.py
@@ -11,21 +11,6 @@ class AbstractDishRepository(ABC):
         pass
 
     @abstractmethod
-    async def add_user(
-        self, user_id: int, username: str | None, is_admin: bool = False
-    ) -> None:
-        """Add a user or update their info."""
-        pass
-
-    @abstractmethod
-    async def delete_user(self, user_id: int) -> None:
-        pass
-
-    @abstractmethod
-    async def get_user(self, user_id: int) -> User | None:
-        pass
-
-    @abstractmethod
     async def add_dish(self, name: str) -> None:
         pass
 
@@ -35,6 +20,26 @@ class AbstractDishRepository(ABC):
 
     @abstractmethod
     async def get_dishes(self) -> Sequence[Dish]:
+        pass
+
+
+class AbstractUserRepository(ABC):
+    @abstractmethod
+    async def init_db(self) -> None:
+        pass
+
+    @abstractmethod
+    async def add_user(
+        self, user_id: int, username: str | None, is_admin: bool = False
+    ) -> None:
+        pass
+
+    @abstractmethod
+    async def delete_user(self, user_id: int) -> None:
+        pass
+
+    @abstractmethod
+    async def get_user(self, user_id: int) -> User | None:
         pass
 
     @abstractmethod
